+++ pydidas/core/generic_params/generic_params_data_import.py
@@ -228,4 +228,57 @@ GENERIC_PARAMS_DATA_IMPORT = {
             " Eiger detector writes its data files."
         ),
     },
+    "raw_datatype": {
+        "type": str,
+        "default": "float 64 bit",
+        "name": "Datatype",
+        "choices": (
+            ["boolean (1 bit integer)"]
+            + [f"float {_i} bit" for _i in [16, 32, 64, 128]]
+            + [f"int {_i} bit" for _i in [8, 16, 32, 64]]
+            + [f" unsigned int {_i} bit" for _i in [8, 16, 32, 64]]
+        ),
+        "unit": "",
+        "allow_None": False,
+        "tooltip": (
+            "The data type to be used for decoding. Note that numpy data types are "
+            "used for decoding with native byteorder."
+        ),
+    },
+    "raw_shape_x": {
+        "type": int,
+        "default": 0,
+        "name": "Raw shape x",
+        "choices": None,
+        "unit": "px",
+        "allow_None": False,
+        "tooltip": (
+            "The x shape of the raw data file. Following the python convention,"
+            "the first axis is y and the second axis is x."
+        ),
+    },
+    "raw_shape_y": {
+        "type": int,
+        "default": 0,
+        "name": "Raw shape y",
+        "choices": None,
+        "unit": "px",
+        "allow_None": False,
+        "tooltip": (
+            "The y shape of the raw data file. Following the python convention,"
+            "the first axis is y and the second axis is x."
+        ),
+    },
+    "raw_header": {
+        "type": int,
+        "default": 0,
+        "name": "Raw file header length",
+        "choices": None,
+        "unit": "bytes",
+        "allow_None": False,
+        "tooltip": (
+            "The length of the file header in bytes. The header will not be "
+            "decoded as image data."
+        ),
+    },
 }