+++ pydidas/gui/frames/builders/data_browsing_frame_builder.py
@@ -34,7 +34,11 @@ from qtpy import QtWidgets
 from ....core.constants import POLICY_EXP_EXP
 from ....widgets.factory import SquareButton
 from ....widgets.framework import BaseFrame
-from ....widgets.selection import DirectoryExplorer, Hdf5DatasetSelector
+from ....widgets.selection import (
+    DirectoryExplorer,
+    Hdf5DatasetSelector,
+    RawMetadataSelector,
+)
 from ....widgets.silx_plot import PydidasImageView
 
 
@@ -58,10 +62,18 @@ class DataBrowsingFrameBuilder:
             parent_widget="browser",
         )
         frame.create_any_widget(
-            "hdf_dset",
+            "hdf5_dataset_selector",
             Hdf5DatasetSelector,
             gridPos=(3, 0, 1, 1),
             parent_widget="browser",
+            visible=False,
+        )
+        frame.create_any_widget(
+            "raw_metadata_selector",
+            RawMetadataSelector,
+            gridPos=(4, 0, 1, 1),
+            parent_widget="browser",
+            visible=False,
         )
         frame.create_any_widget(
             "but_minimize",
@@ -79,8 +91,6 @@ class DataBrowsingFrameBuilder:
         )
 
         frame._widgets["viewer"] = PydidasImageView()
-
-        frame._widgets["hdf_dset"].register_view_widget(frame._widgets["viewer"])
 
         frame._widgets["splitter"] = QtWidgets.QSplitter()
         frame._widgets["splitter"].setSizePolicy(